\name{covariance}
\alias{covariance}
\title{Covariance}
\usage{
covariance(x, y, ...)
}
\arguments{
  \item{x,}{y numerical vectors}

  \item{...}{expressions evaluated in the context of
  \code{x} and \code{y} and passed to other called
  functions}
}
\description{
This function calculates the covariance of two vectors
}
\examples{
covariance(rnorm(10),rnorm(10))
}
\keyword{covariance}

