\docType{package}
\name{rkteaching_es}
\alias{rkteaching_es}
\alias{rkteaching_es-package}
\title{RKTeaching: Un paquete para la ensañanza de Estadística con RKWard}
\description{
RKTeaching proporciona menús y cuadros de diálogo sencillos
para realizar los prodecimientos estadísticos más
habituales en el análisis de datos. Los cuadros de diálogo
están especialmente diseñados para simplificar la
aplicación de los procedimientos para usuarios que están
aprendiendo Estadística sin necesidad de trabajar con
comandos de R.
}

